Move embed related options from filters to embed settings
