+++ react-ui/src/components/EmbedSettings/AdditionalCategories/AdditionalCategories.tsx
