+++ react-ui/src/components/EmbedSettings/LinkContainer/LinkContainer.tsx
@@ -4,12 +4,15 @@ import { bindActionCreators } from '@reduxjs/toolkit';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAppDispatch } from '../../../hooks/rtkHooks';
-import filterSlice from '../../../redux/slices/filterSlice';
+import embedSettingsSlice from '../../../redux/slices/embedSettingsSlice';
 
 const LinkContainer = () => {
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
-  const { setLinkContainer, setLinkText } = bindActionCreators(filterSlice.actions, dispatch);
+  const { setLinkContainer, setLinkText } = bindActionCreators(
+    embedSettingsSlice.actions,
+    dispatch,
+  );
 
   const handleLinkChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     if (e.target.value.trim().length >= 0) {
