+++ react-ui/src/components/EmbedSettings/OpenInNewWindow/OpenInNewWindow.tsx
@@ -6,25 +6,25 @@ import { bindActionCreators } from '@reduxjs/toolkit';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAppDispatch } from '../../../hooks/rtkHooks';
-import filterSlice from '../../../redux/slices/filterSlice';
+import embedSettingsSlice from '../../../redux/slices/embedSettingsSlice';
 
-const ShowPastEvents = () => {
+const OpenInNewWindow = () => {
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
-  const { setShowPastEvents } = bindActionCreators(filterSlice.actions, dispatch);
+  const { setOpenInNewWindow } = bindActionCreators(embedSettingsSlice.actions, dispatch);
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setShowPastEvents((event.target as HTMLInputElement).value);
+    setOpenInNewWindow((event.target as HTMLInputElement).value);
   };
 
   return (
     <div>
       <FormControl sx={{ padding: 2 }}>
-        <p>{`${t('showPastEvents')}`}</p>
+        <p>{`${t('openEventDetailsInNewWindow')}`}</p>
         <RadioGroup
           row
-          defaultValue="false"
-          name="show-past-events-radio-buttons"
+          defaultValue="true"
+          name="open-in-new-window-radio-buttons"
           onChange={handleChange}
         >
           <FormControlLabel value="true" control={<Radio />} label={`${t('yes')}`} />
@@ -35,4 +35,4 @@ const ShowPastEvents = () => {
   );
 };
 
-export default ShowPastEvents;
+export default OpenInNewWindow;
